data-aggregator/requirements.txt
kafka-python==2.0.1
psycopg2-binary
pytz